refactor(metric_reporter): move parsers and reporters into directories
